added jackson-jaxrs support for Jersey. now jackson is used for for a provider. for the annotations jaxb-api-2.1 and jaxb-impl-2.1 were needed ( jackson wanted them )
added org.codehaus.jackson in the web.xml to every servlet as parameter.

also annotated the beans even more so they have a nice output in json ( jaxb seems to work much better now )

This was all made because of the problem with returning an ArrayList<Object> to JSON would only represent the object as json object, not as json object in an array. For that reason, jackson came in town. Thanks, Jackson!